// src/app/models/product.ts
export interface Product {
  productId:number;
  detailID: number;
  brand: string;
  name: string;
  price: number;
  color: string;
  size: number;
  genderType: string;
  star: number;
  picture1: string;
  picture2: string;
  picture3: string;
  catalogID: number;
  type: string;
  status: string;
}